build fresh path if lookahead expanded the state we were moving towards
By checking if the current 'destination' has been expanded, the decision
strategy automatically takes into account the values backed up from its
successors.  This way we don't keep moving to a state which we found out
to be worse than we initially believed.  This is still fine for
completeness, since we're just commiting to a new path to a frontier
node.