working version of new data gathering algorithm.
Although I've only tried it with lsslrta* so far, the new approach
should be able to use any SearchEngine as an initial search provided it
implements get_expanded_states().  The initial search is simply called in
initialize().  We then get its set of expanded states (with a move, to be
precise) and solve it optimally until we're done or out of time.  If the
initial search times out already, we get nothing of course.