implemented naive time bound, that just limits lookahead to a number of ms.
Obviously this isn't correct yet, as it just stops the lookahead after
the given time limit, but it doesn't return the action. But it's a
start.  The easiest way to make it truly real-time would be to
preallocate like a 10th of the allowed time for learning and action
selection.  Then, to make things work even more smoothly, I could do the
following things:
1. Preload all distributions instead of loading them lazily.
2. Track the time it takes to do one lookahead iteration.  Adjust
online how much time can safely be spent in lookahead.
3. First do action selection (learning is only needed for the next
iteration).  Then do the learning.
4. Make learning interruptible (steppable), and if it doesn't finish
within the time limit, we can still return the action within the limit,
finish learning the next iteration, and remember to allot more time for
it next time.