Worked on timing implementation.
Maybe I should start making smaller commits.  In this commit I
reorganized the learning methods.  I call them backup methods now.  They
now have a consistent interface again , and they both inherit from a
Learning base struct.  They also need to be initialized now, and they
need to be called in steps.  This way I can interrupt them if the time
is up, and (using the initialized state from the previous iteration)
continue with whatever didn't finish last time.  A trivial but somewhat
significant change is the reordering of the real_time step method.  We
first select the action and then do the learning.  This ensure that we
always have selected the action early enough, and just spend the
remaining time learning.  So we're guaranteed to stay under the time
limit (unless something goes completely wrong).  I also implemented a
way to dynamically reallocate execution time between lookahead and
learning.  This needs some tuning, but it works in principle.  90/10
lookahead/learning seems very generous, so if this actually works, I can
squeeze out the last few ms to do as much lookahead as possible.  What's
missing is the step in the beginning of the lookahead phase where I
check if I need to finish the learning from the previous iteration
first.  This shouldn't be too hard, but I need to make sure that it
doesn't interfere with the timing adjustments later.  Obviously, if I
have to finish the learning first, I may not be doing as much lookahead,
and I won't need to do as much learning in that iteration as a
consequence.  I just have to establish some kind of dynamic equilibrium.